+++ test/e2e/addon.js
@@ -18,14 +18,14 @@ describe("node native addon", async () => {
   let nwOptions = {
     srcDir: "test/fixture/gyp",
     mode: "build",
-    version: "0.81.0",
+    version: "0.82.0",
     flavor: "sdk",
     platform: getPlatform(platform),
     arch: getArch(arch),
     outDir: "test/fixture/out/gyp",
     cacheDir: "test/fixture/cache",
     glob: false,
-    nativeAddon: true,
+    nativeAddon: "gyp",
   };
 
   it("builds native addon and executes", async () => {