+++ test/fixture/demo.js
@@ -1,8 +1,9 @@
 import nwbuild from "nw-builder";
 
 await nwbuild({
-  srcDir: "app",
+  srcDir: "gyp",
   mode: "build",
   glob: false,
   logLevel: "debug",
+  nativeAddon: "gyp",
 });