+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [4.4.2-beta.4] - 2023-11-03
+
+### Changed
+
+- Use `yauzl-promise` to decompress MacOS build on MacOS platform.
+
+### Removed
+
+- Native package `unzip` usage.
+
 ## [4.4.2-beta.3] - 2023-10-23
 
 ### Added